# src/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Tech Challenge 7SOAT",
    "description": "API documentation for the Tech Challenge 7SOAT",
    "version": "1.0.0"
  },
  "host": "127.0.0.1:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/customer": {
      "get": {
        "description": "Retrieve a list of customers",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "A list of customers"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      },
      "post": {
        "description": "Create a new customer",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "any"
                },
                "taxpayerRegistry": {
                  "type": "string",
                  "example": "any"
                },
                "email": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Customer created successfully"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/customer/{id}": {
      "get": {
        "description": "Retrieve details of a customer by ID",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "ID of the customer to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer details"
          },
          "404": {
            "description": "Customer not found"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      },
      "delete": {
        "description": "Delete a customer by ID",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the customer to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Customer deleted successfully"
          },
          "404": {
            "description": "Customer not found"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/customer/check-if-exists/{taxpayerRegistry}": {
      "get": {
        "description": "Check if a customer exists by taxpayer registry",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "taxpayerRegistry",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Taxpayer registry of the customer to check"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer exists"
          },
          "404": {
            "description": "Customer not found"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/category": {
      "get": {
        "description": "Retrieve a list of categories",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "A list of categories"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/product": {
      "get": {
        "description": "Retrieve a list of products",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "A list of products"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      },
      "post": {
        "description": "Create a new product",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Coke 600ml"
                },
                "price": {
                  "type": "number",
                  "example": 10.0
                },
                "description": {
                  "type": "string",
                  "example": "Product Description"
                },
                "pictures": {
                  "type": "string",
                  "example": "URL"
                },
                "categoryId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Product created successfully"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/product/{id}": {
      "get": {
        "description": "Retrieve details of a product by ID",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "ID of the product to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Product details"
          },
          "404": {
            "description": "Product not found"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      },
      "delete": {
        "description": "Delete a product by ID",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "ID of the product to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted successfully"
          },
          "404": {
            "description": "Product not found"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/order": {
      "get": {
        "description": "Retrieve a list of orders",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "A list of orders"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      },
      "post": {
        "description": "Create a new order",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "customerId": {
                  "type": "string",
                  "example": "any"
                },
                "notes": {
                  "type": "string",
                  "example": "any"
                },
                "productsAndQuantity": {
                  "type": "object",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Order created successfully"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/order/{id}": {
      "get": {
        "description": "Retrieve details of an order by ID",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "ID of the order to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Order details"
          },
          "404": {
            "description": "Order not found"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      },
      "delete": {
        "description": "Delete an order by ID",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the order to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Order deleted successfully"
          },
          "404": {
            "description": "Order not found"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    }
  }
}